@@ -2,18 +2,15 @@ package chronosacaria.mcsa.registry;
 
 import chronosacaria.mcsa.configs.McsaConfig;
 import chronosacaria.mcsa.items.ArmorSets;
-import chronosacaria.mcsa.registry.ArmorsRegistry;
-import net.fabricmc.fabric.api.loot.v1.FabricLootPoolBuilder;
-import net.fabricmc.fabric.api.loot.v1.event.LootTableLoadingCallback;
-import net.minecraft.loot.BinomialLootTableRange;
+import net.fabricmc.fabric.api.loot.v2.LootTableEvents;
+import net.minecraft.loot.LootPool;
 import net.minecraft.loot.LootTables;
 import net.minecraft.loot.entry.ItemEntry;
+import net.minecraft.loot.provider.number.BinomialLootNumberProvider;
 import net.minecraft.util.Identifier;
 
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashSet;
 
 public class LootRegistry {
 
@@ -22,14 +19,10 @@ public class LootRegistry {
     );
 
     public static final Collection<Identifier> STRONGHOLD_LOOT_TABLES =
-            Collections.unmodifiableSet(new HashSet<>(Arrays.asList(
-                LootTables.STRONGHOLD_CORRIDOR_CHEST, LootTables.STRONGHOLD_CROSSING_CHEST, LootTables.STRONGHOLD_LIBRARY_CHEST
-            )));
+            java.util.Set.of(LootTables.STRONGHOLD_CORRIDOR_CHEST, LootTables.STRONGHOLD_CROSSING_CHEST, LootTables.STRONGHOLD_LIBRARY_CHEST);
 
     public static final Collection<Identifier> TEMPLE_LOOT_TABLES =
-            Collections.unmodifiableSet(new HashSet<>(Arrays.asList(
-                    LootTables.JUNGLE_TEMPLE_CHEST, LootTables.JUNGLE_TEMPLE_DISPENSER_CHEST, LootTables.DESERT_PYRAMID_CHEST
-            )));
+            java.util.Set.of(LootTables.JUNGLE_TEMPLE_CHEST, LootTables.JUNGLE_TEMPLE_DISPENSER_CHEST, LootTables.DESERT_PYRAMID_CHEST);
 
     public static final Collection<Identifier> ARMORER_GIFT_LOOT_TABLE = Collections.singletonList(
             LootTables.HERO_OF_THE_VILLAGE_ARMORER_GIFT_GAMEPLAY
@@ -72,11 +65,11 @@ public class LootRegistry {
     }
 
     public static void init() {
-        LootTableLoadingCallback.EVENT.register((((resourceManager, manager, id, supplier, setter) -> {
+        LootTableEvents.MODIFY.register(((resourceManager, lootManager, id, tableBuilder, source) -> {
             if (!id.getNamespace().equals("minecraft"))
                 return;
 
-            FabricLootPoolBuilder poolBuilder;
+            LootPool.Builder lootPoolBuilder;
 
 
             /* * * * * * * * * * * * * * * |
@@ -84,17 +77,17 @@ public class LootRegistry {
             | * * * * * * * * * * * * * * */
 
             if (VILLAGE_SMITH_LOOT_TABLE.contains(id)){
-                poolBuilder = FabricLootPoolBuilder.builder();
-                addArmorSet(poolBuilder, ArmorSets.ELLEGAARD,   McsaConfig.config.getCharacterArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.GABRIEL,     McsaConfig.config.getCharacterArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.IVOR,        McsaConfig.config.getCharacterArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.LUKAS,       McsaConfig.config.getCharacterArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.MAGNUS,      McsaConfig.config.getCharacterArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.OLIVIA,      McsaConfig.config.getCharacterArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.PETRA,       McsaConfig.config.getCharacterArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.SOREN,       McsaConfig.config.getCharacterArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.TIMS_ARMOUR, McsaConfig.config.getCharacterArmourSpawnRate());
-                supplier.pool(poolBuilder);
+                lootPoolBuilder = LootPool.builder();
+                addArmorSet(lootPoolBuilder, ArmorSets.ELLEGAARD,   McsaConfig.config.getCharacterArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.GABRIEL,     McsaConfig.config.getCharacterArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.IVOR,        McsaConfig.config.getCharacterArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.LUKAS,       McsaConfig.config.getCharacterArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.MAGNUS,      McsaConfig.config.getCharacterArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.OLIVIA,      McsaConfig.config.getCharacterArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.PETRA,       McsaConfig.config.getCharacterArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.SOREN,       McsaConfig.config.getCharacterArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.TIMS_ARMOUR, McsaConfig.config.getCharacterArmourSpawnRate());
+                tableBuilder.pool(lootPoolBuilder);
             }
 
             /* * * * * * * * * * * * *|
@@ -102,13 +95,13 @@ public class LootRegistry {
             |* * * * * * * * * * * * */
 
             else if (STRONGHOLD_LOOT_TABLES.contains(id)){
-                poolBuilder = FabricLootPoolBuilder.builder();
-                addArmorSet(poolBuilder, ArmorSets.DRAGONSBANE,             McsaConfig.config.getStrongholdArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.ENDER_DEFENDER,          McsaConfig.config.getStrongholdArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.GOLDEN_GOLIATH,          McsaConfig.config.getStrongholdArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.GOLDEN_GOLIATH_CIRCUITRY,McsaConfig.config.getStrongholdArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.REDSTONE_RIOT,           McsaConfig.config.getStrongholdArmourSpawnRate());
-                supplier.pool(poolBuilder);
+                lootPoolBuilder = LootPool.builder();
+                addArmorSet(lootPoolBuilder, ArmorSets.DRAGONSBANE,             McsaConfig.config.getStrongholdArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.ENDER_DEFENDER,          McsaConfig.config.getStrongholdArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.GOLDEN_GOLIATH,          McsaConfig.config.getStrongholdArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.GOLDEN_GOLIATH_CIRCUITRY,McsaConfig.config.getStrongholdArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.REDSTONE_RIOT,           McsaConfig.config.getStrongholdArmourSpawnRate());
+                tableBuilder.pool(lootPoolBuilder);
             }
 
             /* * * * * * * * * * *|
@@ -116,11 +109,11 @@ public class LootRegistry {
             |* * * * * * * * * * */
 
             else if (TEMPLE_LOOT_TABLES.contains(id)) {
-                poolBuilder = FabricLootPoolBuilder.builder();
-                addArmorSet(poolBuilder, ArmorSets.SHIELD_OF_INFINITY,  McsaConfig.config.getTempleArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.STAR_SHIELD,         McsaConfig.config.getTempleArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.SWORDBREAKER,        McsaConfig.config.getTempleArmourSpawnRate());
-                supplier.pool(poolBuilder);
+                lootPoolBuilder = LootPool.builder();
+                addArmorSet(lootPoolBuilder, ArmorSets.SHIELD_OF_INFINITY,  McsaConfig.config.getTempleArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.STAR_SHIELD,         McsaConfig.config.getTempleArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.SWORDBREAKER,        McsaConfig.config.getTempleArmourSpawnRate());
+                tableBuilder.pool(lootPoolBuilder);
             }
 
             /* * * * * * * * * * * * *|
@@ -128,17 +121,17 @@ public class LootRegistry {
             |* * * * * * * * * * * * */
 
             else if (ARMORER_GIFT_LOOT_TABLE.contains(id)) {
-                poolBuilder = FabricLootPoolBuilder.builder();
-                addArmorSet(poolBuilder, ArmorSets.ADAMANTIUM,      McsaConfig.config.getHovArmourSpawnRate());
-                addArmorSet(poolBuilder, ArmorSets.CHAMPION_PETRA,  McsaConfig.config.getHovArmourSpawnRate());
-                supplier.pool(poolBuilder);
+                lootPoolBuilder = LootPool.builder();
+                addArmorSet(lootPoolBuilder, ArmorSets.ADAMANTIUM,      McsaConfig.config.getHovArmourSpawnRate());
+                addArmorSet(lootPoolBuilder, ArmorSets.CHAMPION_PETRA,  McsaConfig.config.getHovArmourSpawnRate());
+                tableBuilder.pool(lootPoolBuilder);
             }
-        })));
+        }));
     }
-    public static void addArmorSet(FabricLootPoolBuilder poolBuilder, ArmorSets set, float p){
+    public static void addArmorSet(LootPool.Builder poolBuilder, ArmorSets set, float p){
         ArmorsRegistry.armorItems.get(set).values()
                 .forEach((item -> {
-                    poolBuilder.rolls(new BinomialLootTableRange(1, p));
+                    poolBuilder.rolls(BinomialLootNumberProvider.create(1, p));
                     poolBuilder.with(ItemEntry.builder(item));
                 }));
     }
